pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.17"
dependencies = [
    "bfgbidding>=0.0.21",
    "bfgcardplay>=1.2.7",
    "bfgdealer>=0.0.14",
    "bridgeobjects>=0.1.30",
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "icecream>=2.1.4",
    "pillow>=11.2.1",
    "requests>=2.32.5",
    "termcolor>=3.1.0",
]
